+++ src/components/Chat/index.ts
@@ -1,3 +1,4 @@
 export * from "./ChatArea";
 export * from "./ChatContent";
+export * from "./ChatContentCollapsed";
 export * from "./ChatField";